+++ MediaBrowser.Model/MediaInfo/LiveStreamRequest.cs
@@ -7,7 +7,7 @@ namespace MediaBrowser.Model.MediaInfo
         public string OpenToken { get; set; }
         public string UserId { get; set; }
         public string PlaySessionId { get; set; }
-        public int? MaxStreamingBitrate { get; set; }
+        public long? MaxStreamingBitrate { get; set; }
         public long? StartTimeTicks { get; set; }
         public int? AudioStreamIndex { get; set; }
         public int? SubtitleStreamIndex { get; set; }
